backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.1
python-dotenv==1.0.1
mysql-connector-python==9.0.0
